Fix German translation
A "public license" is a "Veröffentlichsungslizenz" (license to publish,
instead of a license that is public).

Er/sie (he/she) is written in lower case if it is not the person who you
are directly talking to.

An option is a "Möglichkeit". A "Option" is a choice you take, a
"Möglichkeit" is force upon you by a someone else, the fates, whatever.

"Im Falle" is also used for options you have no say in.

"Sowohl" means the particular cases most not all apply, but the license
wants none if the cases to apply.

Greetings and good-byes are punctuated with an exclamation point.